requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
certifi==2024.7.4
channels==4.1.0
charset-normalizer==3.3.2
colorama==0.4.6
comtypes==1.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
Django==5.0.7
facenet-pytorch==2.6.0
fer==22.5.1
ffmpeg==1.4
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.65.4
h5py==3.11.0
huggingface-hub==0.24.5
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
keras==3.4.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.4.0
moviepy==1.0.3
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.2.0
proglog==0.1.10
protobuf==4.25.4
Pygments==2.18.0
pyparsing==3.1.2
pypiwin32==223
python-dateutil==2.9.0.post0
pyttsx3==2.90
pytz==2024.1
pywin32==306
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
safetensors==0.4.4
scikit-image==0.24.0
scipy==1.14.0
setuptools==72.1.0
six==1.16.0
SpeechRecognition==3.10.4
sqlparse==0.5.1
sympy==1.13.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
tifffile==2024.7.24
tokenizers==0.19.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.5
transformers==4.43.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.44.0
wldhx.yadisk-direct==0.0.6
wrapt==1.16.0
